tmp: 砍掉檔案，邏輯都要重想一次
