--- src/controller/user/index.ts
@@ -1,2 +1,0 @@
-export { default as getOrCreateNewTenantToGetToken } from './getOrCreateNewTenantToGetToken';
-export { default as getTenantTokenByTenantId } from './getTenantTokenByTenantId';
