--- src/service/user/index.ts
@@ -1,6 +1,0 @@
-export { default as createTenant } from './createTenant';
-export { default as createTenantAdmin } from './createTenantAdmin';
-export { default as getFirstTenantAdminId } from './getFirstTenantAdminId';
-export { default as getFirstTenantId } from './getFirstTenantId';
-export { default as getSystemAdminToken } from './getSystemAdminToken';
-export { default as getTenantToken } from './getTenantToken';
